+++ src/main/org/firebirdsql/gds/impl/ServiceParameterBufferImp.java
@@ -81,7 +81,14 @@ public class ServiceParameterBufferImp extends ParameterBufferBase implements Se
 
             @Override
             public ArgumentType getIntegerArgumentType(int tag) {
-                return ArgumentType.IntSpb;
+                switch (tag) {
+                case ISCConstants.isc_spb_rpr_commit_trans_64:
+                case ISCConstants.isc_spb_rpr_rollback_trans_64:
+                case ISCConstants.isc_spb_rpr_recover_two_phase_64:
+                    return ArgumentType.BigIntSpb;
+                default:
+                    return ArgumentType.IntSpb;
+                }
             }
         },
         SPB_VERSION_3_ATTACH(ISCConstants.isc_spb_version3) {